+++ resources/views/checkout/steps/payment.blade.php
@@ -1,3 +1,5 @@
+@php($checkoutAgreements = \Rapidez\Core\Models\CheckoutAgreement::all())
+
 <h1 class="font-bold text-4xl mb-5">@lang('Payment method')</h1>
 <form class="bg-highlight p-8 rounded mt-6" v-on:submit.prevent="save(['payment_method'], 4)">
     <div class="my-2 border bg-white p-4 rounded" v-for="(method, index) in checkout.payment_methods">
@@ -18,37 +20,47 @@
         </x-rapidez::radio>
     </div>
 
-    <graphql query="{ checkoutAgreements { agreement_id name checkbox_text content is_html mode } }">
-        <div v-if="data" slot-scope="{ data }" class="mt-5">
-            <div v-for="agreement in data.checkoutAgreements">
-                <x-rapidez::slideover>
-                    <x-slot name="button">
-                        <a class="text-gray-700" href="#" v-on:click.prevent="toggle" v-if="agreement.mode == 'AUTO'">
-                            @{{ agreement.checkbox_text }}
-                        </a>
-                        <div v-else>
-                            <x-rapidez::checkbox
-                                name="agreement_ids[]"
-                                v-bind:value="agreement.agreement_id"
-                                v-model="checkout.agreement_ids"
-                                dusk="agreements"
-                                required
-                            >
-                                <a href="#" v-on:click.prevent="toggle">@{{ agreement.checkbox_text }}</a>
-                            </x-rapidez::checkbox>
-                        </div>
-                    </x-slot>
-
+    @if (count($checkoutAgreements))
+        <div>
+            @foreach ($checkoutAgreements as $agreement)
+                <x-rapidez::slideover position="right" id="{{ $agreement->agreement_id }}_agreement">
                     <x-slot name="title">
-                        @{{ agreement.name }}
+                        {{ $agreement->name }}
                     </x-slot>
 
-                    <div v-if="agreement.is_html" v-html="agreement.content"></div>
-                    <div v-else v-text="agreement.content" class="whitespace-pre-line"></div>
+                    @if ($agreement->is_html)
+                        <div>
+                            {!! $agreement->content !!}
+                        </div>
+                    @else
+                        <div class="whitespace-pre-line">
+                            {{ $agreement->content }}
+                        </div>
+                    @endif
                 </x-rapidez::slideover>
-            </div>
+
+                @if ($agreement->mode == 'AUTO')
+                    <label class="text-gray-700 cursor-pointer underline hover:no-underline" for="{{ $agreement->agreement_id }}_agreement">
+                        {{ $agreement->checkbox_text }}
+                    </label>
+                @else
+                    <div>
+                        <x-rapidez::checkbox
+                            name="agreement_ids[]"
+                            :value="$agreement->agreement_id"
+                            v-model="checkout.agreement_ids"
+                            dusk="agreements"
+                            required
+                        >
+                            <label class="text-gray-700 cursor-pointer underline hover:no-underline" for="{{ $agreement->agreement_id }}_agreement">
+                                {{ $agreement->checkbox_text }}
+                            </label>
+                        </x-rapidez::checkbox>
+                    </div>
+                @endif
+            @endforeach
         </div>
-    </graphql>
+    @endif
 
     <x-rapidez::button type="submit" class="mt-5" dusk="continue">
         @lang('Place order')