+++ src/Models/Scopes/ForCurrentStoreScope.php
@@ -5,6 +5,7 @@ namespace Rapidez\Core\Models\Scopes;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Scope;
+use Illuminate\Support\Facades\DB;
 
 class ForCurrentStoreScope implements Scope
 {
@@ -21,12 +22,12 @@ class ForCurrentStoreScope implements Scope
         $joinTable = $this->joinTable ?: $currentTable . '_store';
         $primaryKey = $model->getKeyName();
 
-        $builder
-            ->leftJoin($joinTable, function ($join) use ($currentTable, $primaryKey, $joinTable) {
-                $join->on($currentTable . '.' . $primaryKey, '=', $joinTable . '.' . $primaryKey);
-            })
+        $storeQuery = DB::table($joinTable)
+            ->where($builder->qualifyColumn($primaryKey), DB::raw($joinTable . '.' . $primaryKey))
             ->whereIn('store_id', [0, config('rapidez.store')])
             ->orderByDesc('store_id')
             ->limit(1);
+
+        $builder->joinLateral($storeQuery, $joinTable);
     }
 }